LPD-89981 Add MessageBus async layer for AI guardrail alert routing
The current process() in AIGuardrailAlertAuditMessageProcessor runs the full notification lookup, severity classification, and dispatch (DB + SMTP) synchronously on the caller's thread. With LPD-89836 having Pedro call AuditRouter.route(...) inline from the AI agent's model listener, every guardrail alert holds the agent execution path on database and email I/O until the dispatch completes. The LPD-89981 acceptance criterion explicitly requires this routing to be async and non-blocking, so the sync implementation breaks the contract as soon as the upstream event starts flowing.

The portal-native primitive for decoupling without an external broker is a MessageBus parallel destination. Refactor follows the canonical 2-component shape used by KaleoAssetDeploymentMessageListener (portal-workflow-kaleo-runtime-impl) and DLDeletionMessagingConfigurator (document-library-service).

AIGuardrailAlertAuditMessageProcessor.process() now only serializes the audit message to JSON and enqueues it on the destination "liferay/ai_incident_routing" via MessageBus.sendMessage(...). It returns immediately, so AuditRouter.route() never blocks on I/O. The OSGi service is still registered against eventTypes=AI_GUARDRAIL_ALERT, so the wiring from the AppSec audit framework is unchanged.

A new AIIncidentRoutingMessageListener under com.liferay.ai.hub.internal.messaging extends BaseMessageListener and receives messages from that destination. doReceive(Message) reconstructs the AuditMessage from the JSON payload with new AuditMessage((String)payload) and runs the existing notification logic — preferences lookup, severity classify, term values build, dispatch via NotificationTypeServiceTracker. All the private helpers that lived on the processor moved to the listener verbatim, including NotificationConstants references.

The destination itself is registered in the same listener class in @Activate via DestinationConfiguration.createParallelDestinationConfiguration(...), mirroring the canonical pattern (KaleoAssetDeploymentMessageListener registers its destination inline the same way). No separate DestinationConfigurator class is needed; BaseMessageListener handles the receive() try-catch and the @Deactivate unregisters the destination service.
